refactor: react-router-dom 클라이언트 라우팅 반영
